requirements.txt
a-cv-imwrite-imread-plus==0.13
a-cv2-easy-resize==0.10
absl-py==2.1.0
africastalking==1.1.2
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.1
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.8.2
cameracv2==0.1
canonical==0.52.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
ckms==0.65.0
click==8.1.7
colorama==0.4.6
comtypes==1.4.1
crcmod==1.7
cryptography==42.0.7
cv2-brightness==0.0.1
cv2-rgb==1.0.0
cv2-stack-images==0.10
cv2_enumerate_cameras==1.1.9
cv2_utils==0.2.0
cv2imageload==1.0.6
cv2module==0.1.2
deepface==0.0.91
DevSqlite3==1.2
Django==5.0.6
dlib @ file:///C:/Users/USER/Desktop/dee/Dlib_Windows_Python3.x-main/dlib-19.24.99-cp312-cp312-win_amd64.whl#sha256=20c62e606ca4c9961305f7be3d03990380d3e6c17f8d27798996e97a73271862
dnspython==2.6.1
dsnparse==0.2.1
email_validator==2.1.1
filelock==3.14.0
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
gast==0.5.4
gdown==5.1.0
google-pasta==0.2.0
grpcio==1.63.0
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
hackyargparser==0.12
headless==1.12.4
httpcore==1.0.5
httpx==0.27.0
idna==3.7
image-analyst==0.2.1
image-analyst-cv2==0.2.0
inflect==7.2.1
isiter==0.10
iso3166==2.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.3.3
keyboard==0.13.5
kthread==0.2.3
kthread-sleep==0.11
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.2.0
mtcnn==0.1.1
mysqlclient==2.2.4
namex==0.0.8
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
phonenumbers==8.13.36
PIL-Tools==1.1.0
PIL_ext==0.1
pillow==10.3.0
Pillow-PIL==0.1.dev0
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycv2==0.0.1
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywinpipe==0.11
PyYAML==6.0.1
requests==2.31.0
retina-face==0.0.17
rich==13.7.1
schema==0.7.7
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlite3py==2.0.5
sqlparse==0.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
tkinter-webcam==1.0.9
tolerant-isinstance==0.10
touchtouch==0.11
tqdm==4.66.4
typeguard==4.2.1
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
unimatrix==1.14.0
urllib3==2.2.1
validators==0.28.1
vidizmo-deepface==0.0.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
